Split schema and resolvers into multiple files
