+++ src/types/item/resolvers.js
@@ -1,0 +1,5 @@
+module.exports = {
+  Item: {
+    // sitelinks are the only thing specific to Items that Properties don't have
+  }
+}
