+++ src/types/item/schema.js
@@ -1,0 +1,19 @@
+const { gql } = require('apollo-server');
+
+const typeDefs = gql`
+  type Item implements FingerprintProvider & StatementsProvider {
+    pageid: Int!
+    ns: Int!
+    title: String!
+    lastrevid: Int!
+    modified: String!
+    type: String!
+    id: String!
+    label(language: String): Label
+    description(language: String): Description
+    claims(propertyId: String): [Claim]
+    # sitelinks: [Sitelink]
+  }
+`;
+
+module.exports = typeDefs;
