+++ src/types/query.schema.js
@@ -1,0 +1,7 @@
+const { gql } = require('apollo-server');
+
+module.exports = gql`
+  type Query {
+    item(id: String!): Item!
+  }
+`;
